backend/requirements.txt
# Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.21.0
django-filter==24.1
django-extensions==3.2.3

# Authentication & Security
djangorestframework-simplejwt==5.3.2
PyJWT==2.8.1
cryptography==41.0.7

# Database
psycopg2-binary==2.9.9
django-db-geventpool==4.0.1

# Async & Celery
celery==5.3.4
redis==5.0.1

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# API Documentation
drf-spectacular==0.27.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
Pillow==10.1.0

# Logging & Monitoring
python-json-logger==2.0.7

# Development
black==23.12.1
flake8==6.1.0
isort==5.13.2
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==21.0.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

